mysql: add delay for dry run mode

+++ utils/dynstream/memory_control_test.go
@@ -178,35 +178,35 @@ func TestShouldPausePath(t *testing.T) {
 	areaMemStat := newAreaMemStat(path.area, mc, settings, nil)
 	path.areaMemStat = areaMemStat
 
-	path.pendingSize.Store(uint32(10))
+	path.pendingSize.Store(int64(10))
 	pause, resume := path.areaMemStat.shouldPausePath(path)
 	require.False(t, pause)
 	require.False(t, resume)
 
-	path.pendingSize.Store(uint32(15))
+	path.pendingSize.Store(int64(15))
 	pause, resume = path.areaMemStat.shouldPausePath(path)
 	require.False(t, pause)
 	require.False(t, resume)
 
-	path.pendingSize.Store(uint32(20))
+	path.pendingSize.Store(int64(20))
 	pause, resume = path.areaMemStat.shouldPausePath(path)
 	require.True(t, pause)
 	path.paused.Store(true)
 	require.False(t, resume)
 
-	path.pendingSize.Store(uint32(15))
+	path.pendingSize.Store(int64(15))
 	pause, resume = path.areaMemStat.shouldPausePath(path)
 	require.False(t, pause)
 	require.False(t, resume)
 
-	path.pendingSize.Store(uint32(9))
+	path.pendingSize.Store(int64(9))
 	pause, resume = path.areaMemStat.shouldPausePath(path)
 	require.False(t, pause)
 
 	require.True(t, resume)
 	path.paused.Store(false)
 
-	path.pendingSize.Store(uint32(15))
+	path.pendingSize.Store(int64(15))
 	pause, resume = path.areaMemStat.shouldPausePath(path)
 	require.False(t, pause)
 	require.False(t, resume)
@@ -362,18 +362,18 @@ func TestUpdatePathPauseState(t *testing.T) {
 	mc.addPathToArea(path, settings, feedbackChan)
 	areaMemStat := path.areaMemStat
 
-	path.pendingSize.Store(uint32(10))
+	path.pendingSize.Store(int64(10))
 	areaMemStat.updatePathPauseState(path)
 	require.False(t, path.paused.Load())
 
-	path.pendingSize.Store(uint32(60))
+	path.pendingSize.Store(int64(60))
 	areaMemStat.updatePathPauseState(path)
 	require.True(t, path.paused.Load())
 	fb := <-feedbackChan
 	require.Equal(t, PausePath, fb.FeedbackType)
 	require.Equal(t, path.area, fb.Area)
 
-	path.pendingSize.Store(uint32(9))
+	path.pendingSize.Store(int64(9))
 	areaMemStat.updatePathPauseState(path)
 	require.True(t, path.paused.Load())
